+++ zhihu-analysis-core/src/main/java/org/care/core/dao/base/BaseDao.java
@@ -1,0 +1,5 @@
+package org.care.core.dao.base;
+
+public interface BaseDao {
+
+}
