+++ zhihu-analysis-core/src/main/java/org/care/core/service/base/CrudService.java
@@ -1,0 +1,58 @@
+package org.care.core.service.base;
+
+import org.care.core.dao.base.BaseEntity;
+import org.care.core.dao.base.CrudDao;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.transaction.annotation.Transactional;
+
+public abstract class CrudService<D extends CrudDao<T>, T extends BaseEntity<T>> {
+	
+	/**
+	 * 持久层对象
+	 */
+	@Autowired
+	protected D dao;
+	
+	/**
+	 * 获取单条数据
+	 * @param id
+	 * @return
+	 */
+	public T get(String id) {
+		return dao.get(id);
+	}
+	
+	/**
+	 * 获取单条数据
+	 * @param entity
+	 * @return
+	 */
+	public T get(T entity) {
+		return dao.get(entity);
+	}
+
+	/**
+	 * 保存数据（插入或更新）
+	 * @param entity
+	 */
+	@Transactional(readOnly = false)
+	public void save(T entity) {
+		if (entity.getIsNewRecord()){
+			entity.preInsert();
+			dao.insert(entity);
+		}else{
+			entity.preUpdate();
+			dao.update(entity);
+		}
+	}
+	
+	/**
+	 * 删除数据
+	 * @param entity
+	 */
+	@Transactional(readOnly = false)
+	public void delete(T entity) {
+		dao.delete(entity);
+	}
+
+}
