+++ zhihu-analysis-core/src/main/java/org/care/spider/zhihu/dao/UrlResultDao.java
@@ -1,0 +1,12 @@
+package org.care.spider.zhihu.dao;
+
+import org.care.core.dao.base.CrudDao;
+import org.care.core.dao.base.MyBatisDao;
+import org.care.spider.zhihu.entity.UrlResult;
+
+/**
+ * Created by van persie on 2017/3/2.
+ */
+@MyBatisDao
+public interface UrlResultDao extends CrudDao<UrlResult> {
+}
