+++ zhihu-analysis-core/src/main/java/org/care/spider/zhihu/service/UrlResultService.java
@@ -1,0 +1,13 @@
+package org.care.spider.zhihu.service;
+
+import org.care.core.service.base.CrudService;
+import org.care.spider.zhihu.dao.UrlResultDao;
+import org.care.spider.zhihu.entity.UrlResult;
+import org.springframework.stereotype.Service;
+
+/**
+ * Created by van persie on 2017/3/2.
+ */
+@Service
+public class UrlResultService extends CrudService<UrlResultDao, UrlResult> {
+}
