+++ zhihu-analysis-core/src/main/java/org/care/spider/zhihu/utils/IdGen.java
@@ -1,0 +1,32 @@
+package org.care.spider.zhihu.utils;
+
+import org.springframework.context.annotation.Lazy;
+import org.springframework.stereotype.Service;
+
+import java.security.SecureRandom;
+import java.util.UUID;
+
+@Service
+@Lazy(false)
+public class IdGen {
+
+	private static SecureRandom random = new SecureRandom();
+	
+	/**
+	 * 封装JDK自带的UUID, 通过Random数字生成, 中间无-分割.
+	 */
+	public static String uuid() {
+		return UUID.randomUUID().toString().replaceAll("-", "");
+	}
+	
+	/**
+	 * 使用SecureRandom随机生成Long. 
+	 */
+	public static long randomLong() {
+		return Math.abs(random.nextLong());
+	}
+
+	public String getNextId() {
+		return IdGen.uuid();
+	}
+}
