--- zhihu-analysis-core/src/main/java/org/care/zhihu/spider/dao/UrlResultDao.java
@@ -1,12 +1,0 @@
-package org.care.zhihu.spider.dao;
-
-import org.care.zhihu.spider.dao.base.CrudDao;
-import org.care.zhihu.spider.dao.base.MyBatisDao;
-import org.care.zhihu.spider.entity.UrlResult;
-
-/**
- * Created by van persie on 2017/3/2.
- */
-@MyBatisDao
-public interface UrlResultDao extends CrudDao<UrlResult> {
-}
