--- zhihu-analysis-core/src/main/java/org/care/zhihu/spider/dao/base/BaseDao.java
@@ -1,5 +1,0 @@
-package org.care.zhihu.spider.dao.base;
-
-public interface BaseDao {
-
-}
