--- zhihu-analysis-core/src/main/java/org/care/zhihu/spider/dao/base/DataEntity.java
@@ -1,84 +1,0 @@
-package org.care.zhihu.spider.dao.base;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import org.care.zhihu.spider.utils.IdGen;
-
-import java.util.Date;
-
-public abstract class DataEntity<T> extends BaseEntity<T> {
-
-	private static final long serialVersionUID = 1L;
-	
-	protected String remarks;	// 备注
-	protected Date createDate;	// 创建日期
-	protected Date updateDate;	// 更新日期
-	protected String delFlag; 	// 删除标记（0：正常；1：删除；2：审核）
-	
-	public DataEntity() {
-		super();
-		this.delFlag = DEL_FLAG_NORMAL;
-	}
-	
-	public DataEntity(String id) {
-		super(id);
-	}
-	
-	/**
-	 * 插入之前执行方法，需要手动调用
-	 */
-	@Override
-	public void preInsert(){
-		// 不限制ID为UUID，调用setIsNewRecord()使用自定义ID
-		if (!this.isNewRecord){
-			setId(IdGen.uuid());
-		}
-		this.updateDate = new Date();
-		this.createDate = this.updateDate;
-	}
-	
-	/**
-	 * 更新之前执行方法，需要手动调用
-	 */
-	@Override
-	public void preUpdate(){
-		this.updateDate = new Date();
-	}
-	
-//	@Length(min=0, max=255)
-	public String getRemarks() {
-		return remarks;
-	}
-
-	public void setRemarks(String remarks) {
-		this.remarks = remarks;
-	}
-	
-	@JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-	public Date getCreateDate() {
-		return createDate;
-	}
-
-	public void setCreateDate(Date createDate) {
-		this.createDate = createDate;
-	}
-
-	@JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-	public Date getUpdateDate() {
-		return updateDate;
-	}
-
-	public void setUpdateDate(Date updateDate) {
-		this.updateDate = updateDate;
-	}
-
-	@JsonIgnore
-	public String getDelFlag() {
-		return delFlag;
-	}
-
-	public void setDelFlag(String delFlag) {
-		this.delFlag = delFlag;
-	}
-
-}
