--- zhihu-analysis-core/src/main/java/org/care/zhihu/spider/entity/UrlResult.java
@@ -1,48 +1,0 @@
-package org.care.zhihu.spider.entity;
-
-import org.care.zhihu.spider.dao.base.BaseEntity;
-import org.care.zhihu.spider.utils.IdGen;
-
-/**
- * Created by van persie on 2017/3/2.
- */
-public class UrlResult extends BaseEntity<UrlResult> {
-
-    private String url;
-    private String success;
-    private String remarks;
-
-    public String getUrl() {
-        return url;
-    }
-
-    public void setUrl(String url) {
-        this.url = url;
-    }
-
-    public String getSuccess() {
-        return success;
-    }
-
-    public void setSuccess(String success) {
-        this.success = success;
-    }
-
-    public String getRemarks() {
-        return remarks;
-    }
-
-    public void setRemarks(String remarks) {
-        this.remarks = remarks;
-    }
-
-    @Override
-    public void preInsert() {
-        this.id = IdGen.uuid();
-    }
-
-    @Override
-    public void preUpdate() {
-
-    }
-}
