--- zhihu-analysis-core/src/main/java/org/care/zhihu/spider/service/UrlResultService.java
@@ -1,13 +1,0 @@
-package org.care.zhihu.spider.service;
-
-import org.care.zhihu.spider.dao.UrlResultDao;
-import org.care.zhihu.spider.entity.UrlResult;
-import org.care.zhihu.spider.service.base.CrudService;
-import org.springframework.stereotype.Service;
-
-/**
- * Created by van persie on 2017/3/2.
- */
-@Service
-public class UrlResultService extends CrudService<UrlResultDao, UrlResult> {
-}
